完善登陆及发布权限

+++ blog-common/src/main/java/top/modelx/blog/common/dao/BlogDao.java
@@ -21,5 +21,5 @@ public interface BlogDao {
 
     List<BlogCategoriesVo> getCategoriesList();
 
-    BlogVo getBlogBy(@Param("blogId")Long blogId);
+    BlogVo getBlogBy(@Param("blogId")Integer blogId);
 }
